Updated utility.js by adding a new function getHumanReadableDueDate()! Updated ViewAssessmentTasks.js by importing getHumanReadableDueDate()!
