+++ FrontEndReact/src/View/Admin/View/ViewAssessmentTask/ViewAssessmentTasks.js
@@ -4,7 +4,7 @@ import CustomDataTable from '../../../Components/CustomDataTable.js';
 import { IconButton } from '@mui/material';
 import EditIcon from '@mui/icons-material/Edit';
 import VisibilityIcon from '@mui/icons-material/Visibility';
-import { formatDueDate } from '../../../../utility.js';
+import { formatDueDate, getHumanReadableDueDate } from '../../../../utility.js';
 
 
 
@@ -56,43 +56,14 @@ class ViewAssessmentTasks extends Component {
                     setCellHeaderProps: () => { return { width:"117px"}},
                     setCellProps: () => { return { width:"117px"} },
                     customBodyRender: (assessment_task_id) => {
-                        let dueDate = assessmentTasksToDueDates[assessment_task_id]["due_date"];
-
-                        let timeZone = assessmentTasksToDueDates[assessment_task_id]["time_zone"];
-
-                        dueDate = dueDate.substring(5);
-
-                        var month = Number(dueDate.substring(0, 2)) - 1;
-
-                        dueDate = dueDate.substring(3);
-
-                        var day = Number(dueDate.substring(0, 2));
-
-                        dueDate = dueDate.substring(3);
-
-                        var hour = Number(dueDate.substring(0, 2));
-
-                        var twelveHourClock = hour < 12 ? "am": "pm";
-
-                        hour = hour > 12 ? (hour % 12) : hour;
-
-                        hour = hour === 0 ? 12 : hour;
-
-                        dueDate = dueDate.substring(3);
-
-                        var minute = Number(dueDate.substring(0, 2));
-
-                        const monthNames = ["Jan","Feb","Mar","Apr","May","Jun","Jul","Aug","Sep","Oct","Nov","Dec"];
-
-                        var minutesString = minute < 10 ? ("0" + minute): minute;
-
-                        var timeString = `${hour}:${minutesString}${twelveHourClock}`;
-
-                        var dueDateString = `${monthNames[month]} ${day} at ${timeString} ${timeZone}`;
+                        let dueDateString = getHumanReadableDueDate(
+                            assessmentTasksToDueDates[assessment_task_id]["due_date"],
+                            assessmentTasksToDueDates[assessment_task_id]["time_zone"]
+                        );
 
                         return(
                             <>
-                                {dueDate && dueDateString ? dueDateString : "N/A"}
+                                {assessmentTasksToDueDates[assessment_task_id]["due_date"] && dueDateString ? dueDateString : "N/A"}
                             </>
                         )
                     }
